Add a "core" property that skips stdlib build
This is a first attempt to make a faster incremental build by
omitting the lib (stdlib) module. Because profiles are applied on
top of the default configuration, I had to trim the default
module set and instead activate the lib modules via a missing
property "core". This allows specifying "-Dcore" to skip building
the lib (stdlib) module.

Because the property will normally be unset, all executions
without "-Dcore" will continue to build stdlib.