Remove weird duplication of kwargs with bogus lvars
All kwargs params `a:` were adding lvar of `a` and `a:`.  This doubled
size of static scopes for lvars and ruined the ability to count kwarg lvars
for the sake of error reporting (see #8344).

My fix is a little gross but the place where these `a:` are coming in is
a bit too much to audit and since Prism replaces this parser in JRuby 10
I just did the quick workaround.