Ensure non-null reference when replacing
If we found a vacated reference above, it could be removed from
the ConcurrentWeakHashMap by the time we get to this point,
resulting in a null return value from computeIfPresent. This new
logic uses compute instead, either replacing the null or vacated
reference with a new one or leaving in place any populated
reference added in parallel somewhere else.

Fixes #8359