Len arg is in/out so remove @Out annotation
Specifying that this is only @Out prevents the referenced length
from being copied to the native buffer. This in turn causes
getsockopt to sometimes get invalid length values, since the
allocate memory may be uninitialized (or if zeroed, zero is an
invalid length as well). Removing @Out makes this an in/out var
and the given length gets copied to native before the getsockopt
call.

Fixes intermittent "EINVAL" errors when calling getsockopt.

Fixes #8339