Opt out of ObjectSpace for FString
ObjectSpace for string is already problematic, but in this case we
have two specialized types of strings that are born frozen before
they are fully initialized. Adding these into ObjectSpace requires
installing a hidden variable, which triggers frozen checks, which
fire causing a FrozenError to try to inspect an incompletely-
initialized object. It's simpler to just say that FString (and
the Debug version) never go into ObjectSpace.