Restore isFrozen short circuit for ObjectSpace
ObjectSpace wants to be able to assign a hidden variable, but
calling checkFrozen here leads to child class implementations
being called that may have other side effects. This is another
place where the frozen checks need to be unified and cleaned up,
but for now I restore the !isFrozen short circuit and revert the
checkFrozen call to avoid this unintended behavior.