Clear callInfo in this block body
Enumerable#to_a does not take keywords but also was not handling
any incoming callInfo bits. As a result an incoming
CALL_KEYWORD_EMPTY bit can get stuck and propagte through to the
next call. This does not process any kwargs because it is
effectively a block with signature like |a, *b| but whether it
should do more than clearing callInfo is unknown.

Fixes jruby/jruby#8382