Also clear here even though no obvious bug
For jruby/jruby#8382 we added calls to resetCallInfo in Enumerable
Java-based blocks for `to_a`, `to_h`, and `tally` because they
could leave the CALL_KEYWORD_EMPTY bit stuck for a subsequent call
and break kwarg processing. The Java block for Enumerable `chunk`
also fails to clear callInfo, but since it immediately invokes
`each` on self it is difficult or impossible to reproduce. each
can't take kwargs without breaking, and clears callInfo when done
so there's no way to observe the stuck bit. However, I believe we
should still resetCallInfo here for sanitary reasons, and some
weird Enumerable consumer in the future may find a way to hit this
improbable case.