Always set parent and name even if not eager
Setting a constant has the side effect of also setting the name
when a class has none. By making JI constants lazy, we also skip
the constant set, so the class remains anonymous.

$ jruby -Xji.eager.constants=false -e 'p java.lang.Runnable.name'
nil

This patch restores parent module and name-setting to the last
phase of Java proxy class initialization, so that the proxy class
gets its real name.