Permute indy invoke paths to reduce uncached stack
Passing all uncached calls through the varargs form simplifies
maintenance, but requires an indirect dispatch for the first
invocation. This adds several frames to the stack for such uncached
calls. The patch here permutes the InvokeSite.invoke methods so
they can directly dispatch to the to-be-cached handle, reducing the
number of stack frames in a backtrace.

This is unlikely to reduce actual stack consumption due to the
unoptimized MethodHandle and LambdaForm between the invokeExact
call and the target method.