Skip extension builds for default gems
Seems to be issues launching these subprocesses via Maven. While
we investigate that, temporarily disable extension building for
default gems.

The problem manifests during the default gem install of the fiddle
gem, which generates a dummy Makefile on JRuby because it uses the
FFI version instead. The launch of extconf.rb exhibits problems in
various environments:

* In CI, it ends up running with the system default Java rather
  than the one configured in JAVA_HOME.
* On local Linux environments, it fails to launch due to a "bar
  file descriptor: /bin/sh" error.

It seems that the subprocess launch is losing environment in some
broken way.