Sort the jump tables based on new values
When deserializing a BSwitch we recalculate the values for the
switch based on the new runtime, which mostly means getting new
Symbol IDs. This was fixed in #8209. However we need to ensure
that the jump table is sorted or the binary search for the switch
will sometimes fail.

This patch sorts the related tables based on the new switch values.

Fixes #8421