Reimplement subclasses with a weak linked list
This is similar to the CRuby subclasses list except for two
key features:

* We use weak references instead of pointer values; the GC vacates
  references for us.
* The linked list structure is immutable and concurrency-safe.

This impl is thread-safe and lock-free due to the immutable linked
list structure.

Adding a new class creates a new head node and atomically
reassigns it into the class. Removing a class finds that element
and vacates its reference. Replacing a class first removes the old
and then adds the new. Traversing is a matter of walking the chain
and omitting vacated references.

Periodically, the list must be rebuilt without dead references.
This is hardcoded currently to be when the list contains more than
25% vacated references.

Adding a class is an O(1) operation. Removal, replacement, and
traversal are amortized O(n).

This structure is also lighter-weight than either the original
ConcurrentWeakHashMap or any implementation of WeakHashMap
provided by the JDK, plus it has no lock overhead and very little
synchronization overhead.