Make Time load its Ruby-impl side
This one is confusing.  We have some Ruby impl but time.rb depends
on date.rb which is a stdlib loaded thing.  So we need to load it
after enough of JRuby it bootstrapped but we want this loaded
ultimately the same time as the Java impl of Time.  I just
put it at the end of Ruby#init but this is probably not right?