Continue aborting if polling while aborting
If a thread is killed and then quickly interrupted in another way,
it may end up polling for that other event and stop aborting from
the kill. This ensures that once aborted a thread can never stop
aborting by picking up a new interrupt event.

This broke the new test server in test/mri/net/http/utils.rb when
it killed the server and then closed the socket; while aborting
the closed socket could trigger additional exceptions, stopping
the abort. This ensures it will be seen at the next poll and
resumed.