Limit this binding to arity <= 3
We could support more arities here, but it is difficult to predict
whether that might bind methods that should not be bound (e.g. if
a multiple-arity method had one common uber method with > 3 args it
would be bound, skipping the annotated arities). In order to land
this without introducing too much complexity, I'm limiting it to
arities <= 3.

There may be some overhead from trying and failing to acquire a
specific-arity target repeatedly. I will be refactoring the indy
binding logic to live close to (or inside) the DynamicMethod types
in order to let them slowly cache all arities with any handle
adaptations necessary. This will eliminate the extra overhead by
failing to find a given arity exactly once.