Fix arity specs for blocks in language
The basic problem was we were not matching MRI in returning 1 for
`|a, **k|` and instead returning -2.  There was some other issue
which occurred where a weird signature is not being marked as rest
and messing up arity for Proc#arity.  I plan on taking a stab at
it in next commit but this fix is more important than that so
this will land even if I cannot figure that out.