Fix Proc#arity properly for lambdas
I had a FIXME about this from another fix for arities and I just
followed MRIs lead here and did it the same way.  I originally had
not because the arity was wrong for execution but right for #arity.
This turned out to be because 3.4 changed how procs spread in the
presence of kwargs (previous commit fixed that).