Rest args dup keyword hash
This fix is ugly but in recv_keyword we will notice that it is not
really a keyword but an element of a rest arg.  In that case we will
dup the kwarg into a normal hash argument.

The ugly part is this is handled while processing kwargs but if I
went to figure out how this worked I would think the logic would be
in recv_rest_arg.  Unfortunately, we no longer have the info available
by that point (and honestly it would be painful to propagate to a later
instruction).