Use same logic as IR for kwarg handling in IO#write
When passing kwargs from a ruby2_keywords method to a core "real
keywords" method, we lose the keywordiness of the incoming hash
and raise an argument error for the arity mismatch. By using the
same logic as IR here, we properly handle the incoming r2k hash.

This same patch could be applied generally to all core methods that
accept keywords, but given the rarity of r2k to core "real
keywords" methods we have chosen to only do targeted fixes. This
reduces the potential impact (versus a new, more general solution)
and the required work (9.4.10 is eagerly awaited). We will explore
more reliable, general improvements to core method kwarg handling
in JRuby 10.

Fixes #8398.