Java 21-friendly avoidance of multiplyExact exceptions
This is the JRuby 10 fix for jruby/jruby#8516, using multiplyHigh
to detect overflow rather than the exception-raising
multiplyExact.

As with JRuby 9.4, this fix gets the 2nd and 3rd cases from
jruby/jruby#8516 into a reasonable performance level, with
subsequent profiles showing the remaining overhead is all within
the implementation of JDK's BigInteger (mostly in mod).