Properly check for overflow from high long
If high == 0 we must confirm the low value does not have the most
significant bit set. If the high bit is set the result is in the
upper range of uint64 and still fits into 64 bits.

If high == -1 we must confirm the low value has the most
significant bit set. If the high bit is not set the result is in
the negated upper range of uint64.