Update jar-dependencies to 0.5.3
In 9.4.10.0 we fixed the behavior of jar-dependencies to avoid
activation until after other gems had booted, allowing jar-deps to
be updated independently. That allowed us to release updates to
jar-deps for users to install and use, but we could not release
before having an upgradable version of JRuby.

This commit updates JRuby to jar-deps 0.5.3 for our 9.4.11.0
release, which should finally end issues with the older jar-deps
and the older ruby-maven and ruby-maven-libs it was bound to.

See jruby/jruby#8593 for an example of continuing issues due to
those older library versions being shipped in 9.4.10.0.