Improve locking guarantees for subclasses maps
The new SubclassesWeakMap properly included read/write locking for
the `put` and `forEach` methods, but unfortunately the `remove`
and `isEmpty` methods were not overridden with locking. This went
unnoticed until after the 9.4.11.0 release, leading to
jruby/jruby#8602.

This patch makes two modifications to ensure proper locking:

* Instead of working with the full Map interface, and possibly
  missing other method uses in the future, this narrows the
  subclasses map interface to a RubyClassSet with only unique
  methods that we use. This guarantees only these methods can be
  called, and we can ensure they are implemented properly.
* Add the missing locking to the `remove` method, now named
  `removeClass`.

A subsequent commit will include a test of high-concurrency reading
and writing to these maps which fails prior to this commit and
passes afterwards.

Fixes jruby/jruby#8602.