Avoid creating baseCallSites until needed
These are only used for one method cache: initialize. This method
is only used for constructing new objects of this class's type, so
this is wasted overhead for everything except concrete classes.

The patch here initializes it lazily since many (most?) RubyModule
subclasses will never need it. We don't bother making it volatile
or atomic since there's no interesting state (occasional double
init will hurt nothing).

Worth pointing out that with indy `new` logic, we may also never
need this cache. The first call may be slow path in indy sites,
but so is the first init of this cache and the first call to use
it. A future commit may remove this altogether.