Drop all extraneous args passed to attr
The delegate library introduces a peculiar pattern for calling its
__getobj__ method that passes a block. When this pattern is used
to call an attr_reader, we error because the attr-binding logic did
not drop the ignored block.

The patch here uses SmartBinder to filter to only the self object
(in the case of an attr_reader) and the given single argument (in
the case of an attr_writer), dropping the block and any other
unused arguments.