Use AutoCreateSharedArchive to autogenerate CDS file
This makes the launcher always use AppCDS for improved startup,
automatically regenerating the file if it is incompatible with the
current JDK or JRuby.

This largely addresses the challenges in jruby/jruby#8610, but it
it only supported on Java 21+. This also may not work properly at
all in environments where JRuby's install dir is not writable by
the user.