Remove upfront fillNill in Array
fillNil does a bunch of extra nil sets up to the capacity of the array
whether those values ever end up getting used.

The rationale for this fix is anything >realSize should get nil'd only
after we change realSize.  For appends this never happens.  The case
where it does happen is in aset.  In this case we nil fill between
oldRealSize and one less than new realSize.

From a performance analysis all arrays initially benefit from not pre-
filling nil up to capacity (which is 16 by default).  For arrays which
are doing arbitrary asets past the end of the array it could end up
doing more fills but asets like this are pretty uncommon.  Most of the
time I do not even think it would be doing more fills for this either.