Also tag this behavior for dedup and doc better
In JRuby, we represent frozen, deduplicated strings as an FString
subclass of RubyString. Normal frozen strings will never be stored
in the fstring cache, and cannot be trivially converted to an
FString, so dedup or -@ of a simple frozen string will never
return that string.

There's also some challenging race issues that would have to be
conquered before accepting a simple frozen string into the cache.
Given the edge-case nature of this feature, we do not implement it.

I believe the spec should be removed so I have filed an issue.

See https://github.com/ruby/spec/issues/1249