Raise Java errors during load as RuntimeError
We have long re-raised Java exceptions during load as LoadError,
but this indicates that the problem was with booting up Ruby code
or libraries rather than a more fatal Java error. Exceptions
raised during boot should not be shrugged off as LoadError, and
should instead be raised as the more fatal RuntimeError and dealt
with.

Also, because we raise LoadError, RubyGems will think that it
might be able to help by re-activating the gem in question and
re-requiring, which can lead to confusing double loads and
duplicate constant definitions. By raising

Relates to duplicate constant definitions during load of gem
sources as described in the comment below:

https://github.com/sparklemotion/nokogiri/issues/3452#issuecomment-2695804290