Propagate Java throwables during load
We have long re-raised Java exceptions during load as LoadError,
but this indicates that the problem was with booting up Ruby code
or libraries rather than a more fatal Java error.

Also, because we raise LoadError, RubyGems will think that it
might be able to help by re-activating the gem in question and
re-requiring, which can lead to confusing double loads and
duplicate constant definitions.

We considered wrapping with a different exception, like
RuntimeError, but that doesn't fix the problem of hiding the
original Java throwable in a non-obvious way. Users can specify
the -d flag to print out throwable stack traces during load, but
most don't know to do it and it is often inconvenient.

This patch makes the following changes for Java throwables raised
during load/require:

* If the throwable is an IOException, propagate it as the
  appropriate Ruby error, as we do for other IOException
  propagating out of Java code.
* All other non-Ruby Java throwables get propagated as-is without
  any wrapper.
* Clean up some redundant cathes that are unnecessary now:
  * MainExitException is Unrescuable and will propagate.
  * CatchThrow ditto

Relates to duplicate constant definitions during load of gem
sources as described in the comment below:

https://github.com/sparklemotion/nokogiri/issues/3452#issuecomment-2695804290