Add Marshal support for Data
Data is a confusing type in CRuby, because although it does not
extend from Struct, it looks like Struct internally for all other
purposes.

In JRuby, where we are implementing Data as a specialized
RubyObject subclass, we have no way to identify the instances as
either Data or Struct. This makes marshal's narrowing of Data to
Struct a bit tricky to support.

This issue also means we cannot currently support extending a
Data class with another concrete Ruby class, because the field
specialization must live on the nearest concrete type. The effect
of this is that extended Data classes will fail to see their own
members and raise an error on construction.

Barring the issues with extended Data classes, all specs for Data
marshaling are now green.