Bail out of typed logic if not floatable here
If we get to the float logic and a value under iteration is not
a Float, fixnum, bignum, or Rational, we revert back to slow logic
with a Float object as the current sum. This prevents a Complex
at this point from being treated as floatable and losing its
imaginary part.