Completely disable ErrorHighlight
ErrorHighlight makes extensive use of CRuby's RubyVM namespace and
directly consumes ISeq instruction sequences from the YARV VM. It
does not currently support JRuby, and interferes with error
reporting when loaded. This patch fully disables ErrorHighlight,
only printing a warning if explicitly enabled.

If it's possible to add equivalent JRuby support in the future, we
can re-enable it.

Fixes #8709