Disable automatic CDS features on Java 12 and earlier
Java 13 introduces the "at exit" flag. Prior to that, creating a CDS archive requires a two-step dance of generating a class list and then using that list to create the archive. To make this feature simpler and focus on the JDK levels that support it the way we want, we'll limit it to Java 13+.