Exclude tests for method visibility changes
When JRuby assigns visibility in a child to a method from a parent
we wrap the original method with a visibility-carrying wrapper
method. Unfortunately if the original changes, this wrapper is not
updated. We need to redispatch in the visibility wrapper to find
the real method each time.

This is a long-term issue so I'm excluding these tests with this
commit and note for when we circle back.