Don't attempt to forward block to non-zsuper
Not sure how this has not been noticed up to now, but we should not
be forwarding the passed-in block to super when it is a normal
non-z super. This patch removes that logic and only passes along
any block that was directly provided to the super call.

Fixes jruby/jruby#8728