Remove this warning to match standard set.rb
The standard set library does not warn along any path if an unused
block is passed to initialize. We should do the same until the
maintainers of the set gem decide that the warning should be added.

Fixes jruby/jruby#8728

See ruby/set#41 for a report about the "missing" warning.