JRuby runs ensures and will continue to do so
JRuby handles ensure blocks with Java finally, which means we
cannot opt out of executing them without checking for some thread-
local "skip ensures" flag. But even with that, we do not believe
it is a good practice to allow killing fibers without running
ensure blocks, since ensure blocks are typically put in place to
prevent leaving a scope without cleaning up some resource.

This spec only passed randomly because it has thread races and
sometimes it appeared that the ensures did not run because they
had not yet run by the time the thread itself terminates. I am
tagging it because it was a false positive in any case, and we do
not plan to support this behavior currently.