For unlinking now to eliminate spurious failure
The currently-included version of tempfile outputs a "removing..."
message when $DEBUG is true and a file is left to finalize and
unlink lazily. These specs never cleaned up the temp file properly,
and as a result the $DEBUG mode spec will sometimes capture that
message.

The change here forces the files to be cleaned up immediately by
passing "true" for "unlink_now", which should avoid emitting the
lazy removal warning and make the specs more reliable.