Check for existence of release file, with fallback to 8
Many Java 1.8 builds do not include a `release` file, and other
post-1.8 JDKs may not include the file (though we have not found
any recently). Since the release file started to show up around
Java 9 timeframe, this change assumes if it is not present, we
are probably on Java 8. This is the second change necessary to
make the new launcher script work on older JRuby versions that
still support Java 8.