Add JRubyField annotation for binding Java fields
This hides the plumbing of defining an internal or instance
variable accessor that binds directly to a Java instance field. In
NoMethodError, it is used to bind the "args" field so that it is
seen by Marshal.dump. In other classes, it can be used to avoid
the overhead of dynamically retrieving instance variables when a
field would fit the code better.