Use a normal ClassValue for all such cases
JRuby includes two versions of ClassValue: one with a hard-
referenced Map from Class to value, and one based on a normal
ClassValue but with the added oddity that it hard-references its
values while storing them in the ClassValue slot weakly. This was
originally intended to allow the values to collect properly when
the ClassValue collected, but on recent testing it seems to just
leak in a different way.

This change siwtches everything to normal ClassValue. Given this
change, the example cases in #8842 no longer leak, nor does a more
aggressive torture test that also spins up and tears down JRuby
runtimes as well as interface implementations. A 30-minute run of
that test uses a lot of memory and GCs heavily, but does level off
in both heap and metaspace. Forced GCs reduce both nearly back to
startup levels.