Switch default JI class values to STABLE
After some discussion we decided that eliminating the leak that
was introduced in 9.4.0.0 was more important than maintaining the
leaky behavior, even though only one user has reported an issue.
The risk was deemed minimal; the only behavioral change is that
the hidden, transient InterfaceImpl classes attached to duck-typed
objects will now eventually be unloaded and collected. Since these
classes were being generated fresh for each new object, there's no
impact to reuse or reloading. If instance of these classes are
memoized by some Java code, they will continue to exist as before.
Users that have concerns about running with the new "stable" JI
class values can revert to the old behavior by passing a flag to
JRuby (or setting the equivalent JVM property):

  -Xji.class.values=HARD_MAP

The new behavior is more stable and no longer leaks classes, and
we believe it is the better experience for JRuby users.

See #8842 and #8844