Fix locking of stable class values for JI proxies
Because the proxy calculation tends to recurse for classes, there
may be multiple StableValue preparing proxies at the same time.
Depending on the ordering of the classes encountered, this can
lead to deadlocks.

This fixes the deadlock by matching MapBasedClassValue's one lock
per ClassValue instance, and also properly double-checks the value
which was missing in the previous version.