Clean up some locking around hierarchy modification
There are too many locks in this logic, leading to deadlock
potential as reported in #8846. This commit tries to centralize
most hierarchy-modifying or hierarchy-walking locks to using the
global runtime.getHierarchyLock, and localizes parts of the same
code that are only modifying the current class.

In the future we should consider the thread-safety of the various
structures that make up RubyModule/RubyClass, since currently most
of those structures are themselves thread-safe or lock-guarded, so
much of the extra locking here may be unnecessary. We also have
opportunities to improve some "compute if absent" scenarios and
refine the way we lock for hierarchy walking to be more localized.

Fixes #8846