Embed kwargs hash directly in call operands
In order to go further with directly passing keyword arguments on
the call stack, we need call instructions to aggregate information
on the kwargs structure. Previous to this commit, the kwargs hash
was constructed as an entirely separate operand, copied to the
call's parameters via a temporary variable. This effectively hid
the actual hash structure from logic reading the call object.

The patch here embeds the operand for the Hash directly into the
call's argument list, allowing for it to be processed at the same
time as the call. The keys and values are still processed as their
own operands at call time, and any elaborate hash or argument list
construction pulls the hash back out to its own operand. The change
here only allows simple kwargs to be directly associated with the
call that will use them.

The IR output illustrates the change more clearly.

For a call like "foo(1, bar: 1, baz: 2, &boo)", here's old IR:

  02:  %v_1 := call_0o(self<%self>, callType: VARIABLE, name: baz, potentiallyRefined: false, flags: 0)
  03:  %v_2 := call_0o(self<%self>, callType: VARIABLE, name: quux, potentiallyRefined: false, flags: 0)
  04:  %v_3 := copy(hash<sym<a>=>%v_1,sym<b>=>%v_2>)
  05:  %v_4 := call_0o(self<%self>, callType: VARIABLE, name: boo, potentiallyRefined: false, flags: 0)
  06:  %v_0 := call(self<%self>, fix<1>, %v_3, %v_4, callType: FUNCTIONAL, name: foo, potentiallyRefined: false, flags: 2)

And here is the new IR, with the Hash directly in the call's
list of operands:

  2:  %v_1 := call_0o(self<%self>, callType: VARIABLE, name: baz, potentiallyRefined: false, flags: 0)
  3:  %v_2 := call_0o(self<%self>, callType: VARIABLE, name: quux, potentiallyRefined: false, flags: 0)
  4:  %v_3 := call_0o(self<%self>, callType: VARIABLE, name: boo, potentiallyRefined: false, flags: 0)
  5:  %v_0 := call(self<%self>, fix<1>, hash<sym<a>=>%v_1,sym<b>=>%v_2>, %v_3, callType: FUNCTIONAL, name: foo, potentiallyRefined: false, flags: 2)

In both cases, all arguments and all operands to the hash are
accessed and constructed in the same order, but there's no temp
var required to hold the intermediate hash. It is directly
associated with the call.